# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "product-service",
        "description": "Product Swagger Documentation",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
        "/health": {
            "get": {
                "tags": [
                    "healthcheck"
                ],
                "summary": "Health",
                "operationId": "health",
                "responses": {
                    "200": {
                        "description": "Empty response"
                    }
                }
            }
        },
        "/heartbeat": {
            "get": {
                "tags": [
                    "healthcheck"
                ],
                "summary": "Heartbeat",
                "operationId": "heartbeat",
                "responses": {
                    "200": {
                        "description": "Empty response"
                    }
                }
            }
        },
        "/version": {
            "get": {
                "tags": [
                    "healthcheck"
                ],
                "summary": "Version",
                "description": "Returns app version",
                "operationId": "version",
                "responses": {
                    "200": {
                        "description": "Returns app version"
                    }
                }
            }
        },
        "/api/v1/product": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Returns a paginated list of products",
                "operationId": "listProducts",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Not Acceptable"
                    }
                }
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Creates product",
                "operationId": "createProduct",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateProduct"
                },
                "responses": {
                    "201": {
                        "description": "Returns created product"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/product/{productId}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Returns a product",
                "operationId": "readProduct",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of product to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid product"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    }
                }
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Deletes a product",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Product id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success, no content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "products"
                ],
                "summary": "Updates a product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of product that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Validation error"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "415": {
                        "description": "Unsupported Media Type"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ValidationErrors": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "parameter": {
                                "description": "Key: request parameter name, value: array of parameter validation errors.",
                                "type": "array",
                                "items": {
                                    "example": "[The order details field is required.]"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Links": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://product.local:81/v1/product?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://product.local:81/v1/product?page=10"
                    },
                    "prev": {
                        "type": "string",
                        "example": "null"
                    },
                    "next": {
                        "type": "string",
                        "example": "http://product.local:81/v1/product?page=2"
                    }
                },
                "type": "object"
            },
            "Meta": {
                "properties": {
                    "current_page": {
                        "type": "int",
                        "example": "1"
                    },
                    "from": {
                        "type": "int",
                        "example": "1"
                    },
                    "last_page": {
                        "type": "int",
                        "example": "1"
                    },
                    "path": {
                        "type": "string",
                        "example": "http://product.local:81/v1/product"
                    },
                    "per_page": {
                        "type": "int",
                        "example": "10"
                    },
                    "to": {
                        "type": "int",
                        "example": "2"
                    },
                    "total": {
                        "type": "int",
                        "example": "2"
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product model",
                "description": "Product model",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    },
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "type": "number",
                        "format": "float"
                    },
                    "currency": {
                        "title": "Currency",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
            "CreateProduct": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "number"
                                },
                                "currency": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UpdateProduct": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "number"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Bearer token",
                "name": "Authorization",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "healthcheck"
        },
        {
            "name": "products"
        }
    ]
}